% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppSNAP-package.R
\docType{package}
\name{RcppSNAP-package}
\alias{RcppSNAP-package}
\alias{_PACKAGE}
\title{RcppSNAP: 'Rcpp' Integration for the SNAP Network Library}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

Other contributors:
\itemize{
  \item  Stanford SNAP (http://snap.stanford.edu) [contributor]
}

}
